Adjust persistent context tests
