+++ tests/test_browser_contexts.py
@@ -5,7 +5,9 @@ from pathlib import Path
 from uuid import uuid4
 
 import pytest
+from playwright.async_api import Browser, TimeoutError as PlaywrightTimeoutError
 from scrapy import Spider, Request
+from scrapy_playwright.page import PageMethod
 
 from tests import make_handler
 from tests.mockserver import StaticMockServer
@@ -24,6 +26,28 @@ class MixinTestCaseMultipleContexts:
         settings = {"PLAYWRIGHT_BROWSER_TYPE": self.browser_type, "CONCURRENT_REQUESTS": 9876}
         async with make_handler(settings) as handler:
             assert handler.max_pages_per_context == 9876
+
+    @pytest.mark.asyncio
+    async def test_context_kwargs(self):
+        settings_dict = {
+            "PLAYWRIGHT_BROWSER_TYPE": self.browser_type,
+            "PLAYWRIGHT_CONTEXTS": {
+                "default": {"java_script_enabled": False},
+            },
+        }
+        async with make_handler(settings_dict) as handler:
+            with StaticMockServer() as server:
+                req = Request(
+                    url=server.urljoin("/scroll.html"),
+                    meta={
+                        "playwright": True,
+                        "playwright_page_methods": [
+                            PageMethod("wait_for_selector", selector="div.quote", timeout=1000),
+                        ],
+                    },
+                )
+                with pytest.raises(PlaywrightTimeoutError):
+                    await handler._download_request(req, Spider("foo"))
 
     @pytest.mark.asyncio
     async def test_contexts_max_pages(self):
@@ -100,30 +124,46 @@ class MixinTestCaseMultipleContexts:
             assert cookie["domain"] == "example.org"
 
     @pytest.mark.asyncio
-    async def test_persistent_contexts(self):
-        temp_dir_foo = f"{tempfile.gettempdir()}/{uuid4()}"
-        temp_dir_bar = f"{tempfile.gettempdir()}/{uuid4()}"
+    async def test_persistent_context(self):
+        temp_dir = f"{tempfile.gettempdir()}/{uuid4()}"
         settings = {
             "PLAYWRIGHT_BROWSER_TYPE": self.browser_type,
             "PLAYWRIGHT_DEFAULT_NAVIGATION_TIMEOUT": 3000,
             "PLAYWRIGHT_CONTEXTS": {
-                "foo": {
-                    "user_data_dir": temp_dir_foo,
-                },
-                "bar": {
-                    "user_data_dir": temp_dir_bar,
+                "persistent": {
+                    "user_data_dir": temp_dir,
                 },
             },
         }
-        assert not Path(temp_dir_foo).exists()
-        assert not Path(temp_dir_bar).exists()
+        assert not Path(temp_dir).exists()
         async with make_handler(settings) as handler:
-            assert Path(temp_dir_foo).is_dir()
-            assert Path(temp_dir_bar).is_dir()
-            assert len(handler.contexts) == 2
-            assert handler.contexts["foo"].persistent
-            assert handler.contexts["bar"].persistent
+            assert Path(temp_dir).is_dir()
+            assert len(handler.contexts) == 1
+            assert handler.contexts["persistent"].persistent
             assert not hasattr(handler, "browser")
+
+    @pytest.mark.asyncio
+    async def test_mixed_persistent_contexts(self):
+        temp_dir = f"{tempfile.gettempdir()}/{uuid4()}"
+        settings = {
+            "PLAYWRIGHT_BROWSER_TYPE": self.browser_type,
+            "PLAYWRIGHT_DEFAULT_NAVIGATION_TIMEOUT": 3000,
+            "PLAYWRIGHT_CONTEXTS": {
+                "persistent": {
+                    "user_data_dir": temp_dir,
+                },
+                "non-persistent": {
+                    "java_script_enabled": False,
+                },
+            },
+        }
+        assert not Path(temp_dir).exists()
+        async with make_handler(settings) as handler:
+            assert Path(temp_dir).is_dir()
+            assert len(handler.contexts) == 2
+            assert handler.contexts["persistent"].persistent
+            assert not handler.contexts["non-persistent"].persistent
+            assert isinstance(handler.browser, Browser)
 
     @pytest.mark.asyncio
     async def test_contexts_dynamic(self):
